+++ packages/vite-plugin-angular/src/index.ts
@@ -1,5 +1,9 @@
 import { angular } from './lib/angular-vite-plugin.js';
 export { PluginOptions } from './lib/angular-vite-plugin.js';
 export { compileAnalogFile } from './lib/authoring/analog.js';
+export {
+  MarkdownTemplateTransform,
+  defaultMarkdownTemplateTransforms,
+} from './lib/authoring/markdown-transform.js';
 
 export default angular;