+++ packages/vite-plugin-angular/src/lib/angular-vite-plugin.ts
@@ -33,6 +33,10 @@ import { angularVitestPlugin } from './angular-vitest-plugin.js';
 const require = createRequire(import.meta.url);
 
 import { getFrontmatterMetadata } from './authoring/frontmatter.js';
+import {
+  defaultMarkdownTemplateTransforms,
+  MarkdownTemplateTransform,
+} from './authoring/markdown-transform.js';
 
 export interface PluginOptions {
   tsconfig?: string;
@@ -54,6 +58,7 @@ export interface PluginOptions {
       | {
           include: string[];
         };
+    markdownTemplateTransforms?: MarkdownTemplateTransform[];
   };
   supportedBrowsers?: string[];
   transformFilter?: (code: string, id: string) => boolean;
@@ -100,6 +105,9 @@ export function angular(options?: PluginOptions): Plugin[] {
     supportedBrowsers: options?.supportedBrowsers ?? ['safari 15'],
     jit: options?.jit,
     supportAnalogFormat: options?.experimental?.supportAnalogFormat ?? false,
+    markdownTemplateTransforms:
+      options?.experimental?.markdownTemplateTransforms ??
+      defaultMarkdownTemplateTransforms,
   };
 
   // The file emitter created during `onStart` that will be used during the build in `onLoad` callbacks for TS files
@@ -507,6 +515,7 @@ export function angular(options?: PluginOptions): Plugin[] {
         inlineStylesExtension: pluginOptions.inlineStylesExtension,
         supportAnalogFormat: pluginOptions.supportAnalogFormat,
         isProd,
+        markdownTemplateTransforms: pluginOptions.markdownTemplateTransforms,
       });
     }
   }